+++ Modules/Intent.Modules.ModuleBuilder.Java/release-notes.md
@@ -1,3 +1,8 @@
+### Version 3.4.0
+
+- Update: Partial Class Access modifier will now be `public` when `Java File Builder` or `Custom` is used as Templating method.
+- New: File Builder (Builder Pattern paradigm) selectable as Templating Method for Templates.
+
 ### Version 3.3.8
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.